비동기 처리 오류를 고친다
 - @Async 메서드가 자신을 호출하는 메서드와 같은 클래스에 있으면 동기
   방식으로 처리된다
 - 스프링에서 @Async를 프록시 기반 AOP를 이용해서 구현하기 때문에, 다른
   @Async 메서드와 이를 호출하는 메서드가 서로 다른 클래스에 있어야 한다
 - 따라서 두 로직이 서로 다른 클래스에 위치하도록 분리한다
 - https://stackoverflow.com/questions/73258139/why-async-is-not-working-if-i-keep-async-method-in-same-class
